Moves route around to support landing page
