+++ package.json
@@ -31,6 +31,7 @@
     "immutable": "^3.8.2",
     "isomorphic-fetch": "^2.2.1",
     "koa": "^2.4.1",
+    "koa-send": "^4.1.3",
     "next": "^4.1.4",
     "next-routes": "^1.1.0",
     "normalize.css": "^7.0.0",
