+++ routes.js
@@ -2,8 +2,7 @@ const routes = (module.exports = require('next-routes')());
 
 routes
   // routes.add(name, pattern = /name, page = name)
-  .add('home', '/', 'index')
-  .add('articles', '/', 'index')
+  .add('articles', '/articles', 'articles')
   .add('replies', '/replies', 'replies')
   // routes.add({name: 'name', pattern: '/name', page: 'name'})
   .add({
